Move copied generated code in loki exporter to third_party directory
