+++ exporter/lokiexporter/exporter.go
@@ -32,7 +32,7 @@ import (
 	"go.opentelemetry.io/collector/consumer/pdata"
 	"go.uber.org/zap"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/lokiexporter/logproto"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/lokiexporter/third_party/logproto"
 )
 
 type lokiExporter struct {
