+++ exporter/lokiexporter/exporter_test.go
@@ -35,7 +35,7 @@ import (
 	"go.opentelemetry.io/collector/translator/conventions"
 	"go.uber.org/zap"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/lokiexporter/logproto"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/lokiexporter/third_party/logproto"
 )
 
 const (
